Abstract AllReduce trainer
